crews.go

+++ crews.go
@@ -10,7 +10,7 @@ type Crews struct {
 	q float64
 }
 
-func NewCrews(filterSize int, numHashes int, primaryThreshold float64, secondaryThreshold float64, numWebsites uint64, p float64, q float64) *Crews {
+func NewCrews(filterSize int, numHashes int, primaryThreshold float64, secondaryThreshold float64, numWebsites int, p float64, q float64) *Crews {
 	return &Crews{
 		primary: NewBloomFilter(uint(filterSize), uint(numHashes)),
 		secondary:   NewBloomFilter(uint(filterSize), uint(numHashes)),
